azure-financial-copilot/requirements.txt
openai>=1.54.0
pandas>=2.1.0
python-dotenv>=1.0.1
streamlit>=1.29.0
pyyaml>=6.0.1
